fixed batch disk space check

+++ Code/OS/EmbeddedPython.cs
@@ -69,7 +69,7 @@ namespace Cupscale.OS
             await Task.Delay(10);
 
             Print("Checking disk space before installation...");
-            int diskSpaceMb = IOUtils.GetDiskSpace(IOUtils.GetAppDataDir());
+            float diskSpaceMb = IOUtils.GetDiskSpace(IOUtils.GetAppDataDir());
             Print($"Available disk space on the current drive: {diskSpaceMb} MB.");
             if (diskSpaceMb < 5000)
             {
